% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quacs_functions.R
\name{census_api_acs5}
\alias{census_api_acs5}
\alias{census_api_acs5_multi}
\title{Get ACS5 Year Estimate Data}
\usage{
census_api_acs5(
  get,
  geo = "for=tract:*",
  where = "in=state:18",
  key = "key=c8c7ed159026d564cdd274066453f5a8496ff558",
  other = "-",
  year = "2017"
)

census_api_acs5_multi(
  var_names,
  geo = "-",
  where = "-",
  key = "-",
  other = "-",
  year = "2017"
)
}
\arguments{
\item{get}{Character vector of varaibles to get, length < 20}

\item{geo}{Geographic argument to ACS API}

\item{where}{Where argument to ACS API}

\item{key}{Key argument to ACS API}

\item{other}{Other argument to ACS API}

\item{year}{Ending year for the dataset}

\item{var_names}{Character vector of varaibles to get, length >= 20}
}
\value{
Resulting table as a data frame
}
\description{
The API imposes limits on the number of varaibles extracted. To extract more than 20,
`census_api_acs5_multi` batches up the requests and combines them
}
