% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quacs_functions.R
\name{.named}
\alias{.named}
\title{Named}
\usage{
.named(x)
}
\arguments{
\item{x}{A data frame}
}
\value{
a data frame with column names and missingness
}
\description{
Internal function for printing data dictionaries
}
