% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quacs_functions.R
\name{get_acs_dd}
\alias{get_acs_dd}
\title{Get ACS Data Dictionary}
\usage{
get_acs_dd(year = "2018")
}
\arguments{
\item{year}{End year of the 5-Year ACS Period}
}
\value{
The data dictionary as a data frame
}
\description{
Get the ACS Data Dictionary for a given year
}
