% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quacs_functions.R
\name{get_url}
\alias{get_url}
\title{Get URL
Get text from a url.}
\usage{
get_url(url_base, url_path)
}
\arguments{
\item{url_base}{the base url, can be https}

\item{url_path}{the url path}
}
\value{
the text of the webpage
}
\description{
Get URL
Get text from a url.
}
