% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quacs.R, R/quacs_app.R
\docType{package}
\name{quacs}
\alias{quacs}
\title{quacs: A Shiny App For Quick Use of ACS Data}
\usage{
quacs()
}
\value{
The table created by the app
}
\description{
This can be called as as function or run as an addin
}
